Refactor Effects for new BaseValue Attribute system
