+++ Source/GMCAbilitySystem/Public/Effects/GMCAbilityEffect.h
@@ -6,6 +6,7 @@
 #include "GameplayTagContainer.h"
 #include "UObject/Object.h"
 #include "GMCAbilitySystem.h"
+#include "Attributes/GMCAttributeModifier.h"
 #include "GMCAbilityEffect.generated.h"
 
 class UGMC_AbilitySystemComponent;
@@ -24,45 +25,6 @@ enum class EEffectState : uint8
 	Initialized,  // Applies Instantly
 	Started, // Lasts for X time
 	Ended  // Lasts forever
-};
-
-UENUM(BlueprintType)
-enum class EModifierType : uint8
-{
-	// Adds to value
-	Add,
-	// Adds to value multiplier. Base Multiplier is 1. A modifier value of 1 will double the value.
-	Multiply,
-	// Adds to value divisor. Base Divisor is 1. A modifier value of 1 will halve the value.
-	Divide     
-};
-
-USTRUCT(BlueprintType)
-struct FGMCAttributeModifier
-{
-	GENERATED_BODY()
-		
-	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Attribute", meta = (Categories="Attribute"))
-	FGameplayTag AttributeTag;
-
-	// Value to modify the attribute by
-	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite)
-	float Value{0};
-
-	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite)
-	EModifierType ModifierType{EModifierType::Add};
-
-	// Only affects Add modifiers
-	// Should this value directly modify the Attribute's BaseValue
-	// Set True for temporary effects, False for permanent effects
-	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite)
-	bool bApplyToBaseValue = false;
-
-	// Metadata tags to be passed with the attribute
-	// Ie: DamageType (Element.Fire, Element.Electric), DamageSource (Source.Player, Source.Boss), etc
-	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite)
-	FGameplayTagContainer MetaTags;
-	
 };
 
 // Container for exposing the attribute modifier to blueprints
@@ -109,9 +71,7 @@ struct FGMCAbilityEffectData
 	bool bIsInstant = true;
 
 	// Apply an inversed version of the modifiers at effect end
-	// Does not apply to Instant effects
-	// Won't work well for periodic effects or anything beyond simple effect modifications
-	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite)
+	UPROPERTY()
 	bool bNegateEffectAtEnd = true;
 
 	// Delay before the effect starts
